+++ public/sw.js
@@ -5,11 +5,12 @@ const APP_SHELL_URLS = [
   "/en",
   "/bs",
   "/manifest.webmanifest",
-  "/icon.svg",
   "/icon-192.png",
   "/icon-512.png",
   "/icon-maskable-512.png",
   "/apple-touch-icon.png",
+  "/favicon-32.png",
+  "/favicon-48.png",
   "/favicon.ico",
   OFFLINE_FALLBACK_URL,
 ];