+++ src/app/layout.tsx
@@ -28,7 +28,8 @@ export async function generateMetadata(): Promise<Metadata> {
     applicationName: t("applicationName"),
     icons: {
       icon: [
-        { url: "/icon.svg", type: "image/svg+xml" },
+        { url: "/favicon-32.png", sizes: "32x32", type: "image/png" },
+        { url: "/favicon-48.png", sizes: "48x48", type: "image/png" },
         { url: "/favicon.ico", sizes: "48x48" },
       ],
       apple: "/apple-touch-icon.png",